put all examples in a single app, use c++14

+++ src/sdl_texture.cpp
@@ -10,7 +10,7 @@
 SDLTexture::SDLTexture(SDL_Renderer* ren, VideoMemory *vram) : ren(ren), vram(vram) {
   auto buf_size = vram->GetWidth() * vram->GetHeight() * 4;
   // TODO: buf_size should be > 0. Maybe throw an exception?
-  buffer = std::unique_ptr<uint8_t[]>(new uint8_t[buf_size]);
+  buffer = std::unique_ptr<uint8_t[]>(new uint8_t[static_cast<unsigned long>(buf_size)]);
   tex = std::unique_ptr<SDL_Texture, sdl_deleter>(SDL_CreateTexture(ren,
                                                                     SDL_PIXELFORMAT_ABGR8888,
                                                                     SDL_TEXTUREACCESS_STREAMING,
